# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CVCTypeEnum": {
				"enum": [
					"UNKNOWN",
					"CENTRAL",
					"STATE",
					"PRIVATE"
				],
				"type": "string"
			},
			"CVCSiteAddress": {
				"properties": {
					"block": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"pincode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"block",
					"district",
					"state",
					"pincode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CVCOperationTime": {
				"properties": {
					"start_time": {
						"type": "string",
						"description": "Format: HH:MM"
					},
					"end_time": {
						"type": "string",
						"description": "Format: HH:MM"
					}
				},
				"required": [
					"start_time",
					"end_time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeoPoint": {
				"properties": {
					"latitude": {
						"type": "string"
					},
					"longitude": {
						"type": "string"
					}
				},
				"required": [
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CVCStatusEnum": {
				"enum": [
					"UNKNOWN",
					"ACTIVE",
					"CLOSED",
					"OUT_OF_STOCK"
				],
				"type": "string"
			},
			"VaccineTypeEnum": {
				"enum": [
					"COVAXIN",
					"COVISHIELD"
				],
				"type": "string"
			},
			"Vaccine": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/VaccineTypeEnum"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"type",
					"count",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CVCData": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"cowin_center_id": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CVCTypeEnum"
					},
					"address": {
						"$ref": "#/components/schemas/CVCSiteAddress"
					},
					"last_verified_at": {
						"type": "string",
						"format": "date-time"
					},
					"slots": {
						"items": {
							"$ref": "#/components/schemas/CVCOperationTime"
						},
						"type": "array"
					},
					"operation_timings": {
						"$ref": "#/components/schemas/CVCOperationTime"
					},
					"geo": {
						"$ref": "#/components/schemas/GeoPoint"
					},
					"vaccine_count": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/CVCStatusEnum"
					},
					"next_stock_refresh_on": {
						"type": "string",
						"format": "date-time"
					},
					"google_maps_url": {
						"type": "string"
					},
					"vaccines": {
						"items": {
							"$ref": "#/components/schemas/Vaccine"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"cowin_center_id",
					"type",
					"address",
					"last_verified_at",
					"slots",
					"operation_timings",
					"vaccine_count",
					"status",
					"google_maps_url",
					"vaccines"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedCVCData": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page_number": {
						"type": "number",
						"format": "double"
					},
					"page_size": {
						"type": "number",
						"format": "double"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/CVCData"
						},
						"type": "array"
					}
				},
				"required": [
					"total",
					"page_number",
					"page_size",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SortOrderEnum": {
				"enum": [
					"ASC",
					"DESC"
				],
				"type": "string"
			},
			"CVCRequest": {
				"properties": {
					"district": {
						"type": "string",
						"description": "Name of the district.\nEither this or pincode is required"
					},
					"pincode": {
						"type": "number",
						"format": "double",
						"description": "Pincode to search with.\nEither this or district is required"
					},
					"page_number": {
						"type": "number",
						"format": "double",
						"default": "1"
					},
					"page_size": {
						"type": "number",
						"format": "double",
						"default": "25"
					},
					"sort": {
						"properties": {
							"distance": {
								"$ref": "#/components/schemas/SortOrderEnum"
							}
						},
						"required": [
							"distance"
						],
						"type": "object"
					},
					"filter": {
						"properties": {
							"availability": {
								"type": "boolean",
								"default": "true"
							},
							"status": {
								"items": {
									"$ref": "#/components/schemas/CVCStatusEnum"
								},
								"type": "array"
							},
							"radius": {
								"type": "number",
								"format": "double",
								"default": "30"
							},
							"vaccines": {
								"items": {
									"$ref": "#/components/schemas/VaccineTypeEnum"
								},
								"type": "array"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CallRequest": {
				"properties": {
					"pincode": {
						"type": "number",
						"format": "double"
					},
					"cvc_site_id": {
						"type": "string"
					},
					"claim": {
						"type": "boolean",
						"description": "if set tp true, will lock the cvc for 20 minutes"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CallResponseEnum": {
				"enum": [
					"RESPONDED",
					"NOT_RESPONDED",
					"WRONG_NUMBER"
				],
				"type": "string"
			},
			"ReportRequest": {
				"properties": {
					"cvc_site_id": {
						"type": "string"
					},
					"call_response": {
						"$ref": "#/components/schemas/CallResponseEnum"
					},
					"status": {
						"$ref": "#/components/schemas/CVCStatusEnum"
					},
					"cvcType": {
						"$ref": "#/components/schemas/CVCTypeEnum"
					},
					"vaccines": {
						"items": {
							"$ref": "#/components/schemas/Vaccine"
						},
						"type": "array"
					},
					"address": {
						"$ref": "#/components/schemas/CVCSiteAddress"
					},
					"operation_timings": {
						"items": {
							"$ref": "#/components/schemas/CVCOperationTime"
						},
						"type": "array"
					},
					"next_stock_refresh_on": {
						"type": "string",
						"format": "date-time"
					},
					"verification_time": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-server",
		"version": "1.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/cvc": {
			"post": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedCVCData"
								}
							}
						}
					}
				},
				"tags": [
					"CVC"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CVCRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/volunteer/call_request": {
			"post": {
				"operationId": "CallRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Volunteer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CallRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/volunteer/report": {
			"post": {
				"operationId": "Report",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Volunteer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReportRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}